Tests/reorganize: move Auth CRAM MD5 test to own file
As this test is marked _incomplete_, no further review of the test has been done and no `@covers` tag has been added.
