+++ test/PHPMailer/PHPMailerTest.php
@@ -36,32 +36,6 @@ final class PHPMailerTest extends TestCase
     }
 
     /**
-     * Test CRAM-MD5 authentication.
-     * Needs a connection to a server that supports this auth mechanism, so commented out by default.
-     */
-    public function testAuthCRAMMD5()
-    {
-        $this->markTestIncomplete(
-            'Test needs a connection to a server supporting the CRAMMD5 auth mechanism.'
-        );
-
-        $this->Mail->Host = 'hostname';
-        $this->Mail->Port = 587;
-        $this->Mail->SMTPAuth = true;
-        $this->Mail->SMTPSecure = PHPMailer::ENCRYPTION_STARTTLS;
-        $this->Mail->AuthType = 'CRAM-MD5';
-        $this->Mail->Username = 'username';
-        $this->Mail->Password = 'password';
-        $this->Mail->Body = 'Test body';
-        $this->Mail->Subject .= ': Auth CRAM-MD5';
-        $this->Mail->From = 'from@example.com';
-        $this->Mail->Sender = 'from@example.com';
-        $this->Mail->clearAllRecipients();
-        $this->Mail->addAddress('user@example.com');
-        //self::assertTrue($this->mail->send(), $this->mail->ErrorInfo);
-    }
-
-    /**
      * Word-wrap an ASCII message.
      */
     public function testWordWrap()
